Rename LoggingExtensions to EnumerationErrorStrategyExtensions
