+++ Cql/Cql.Runtime/Runtime/ErroredEnumerationContinuation.cs
@@ -1,4 +1,12 @@
-﻿namespace Hl7.Cql.Runtime;
+﻿/*
+ * Copyright (c) 2025, Firely, NCQA and contributors
+ * See the file CONTRIBUTORS for details.
+ *
+ * This file is licensed under the BSD 3-Clause license
+ * available at https://raw.githubusercontent.com/FirelyTeam/firely-cql-sdk/main/LICENSE
+ */
+
+namespace Hl7.Cql.Runtime;
 
 /// <summary>
 /// The exception handling to use while processing a batch of items.
